moved Peers to own JS file, trying to prevent freezing w/ timeout
