+++ app/actions/sync.js
@@ -1,19 +1,16 @@
 const base64 = require('base64-arraybuffer')
-const cuid = require('cuid')
 const debug = require('debug')('peermusic:sync:actions')
-const events = require('events')
-const inherits = require('inherits')
 var coversActions = require('./covers.js')
 var fs = require('file-system')(['image/jpeg', 'image/jpg', '', 'audio/mp3', 'audio/wav', 'audio/ogg'])
 var musicSimilarity = require('music-similarity')
+var Peers = require('./sync/peers')
 
-inherits(Peers, events.EventEmitter)
 var peers
 
 var actions = {
   INITIATE_SYNC: () => {
     return (dispatch, getState) => {
-      peers = new Peers(dispatch, getState)
+      peers = new Peers(dispatch, getState, peers)
 
       peers.on('data', function (data, peerId) {
         actions.PROCESS_INCOMING_DATA(data, peerId)(dispatch, getState)
@@ -627,146 +624,6 @@ var actions = {
         type: 'UPDATED_DEVICE_LIST',
         version: deviceListVersion
       })
-    }
-  }
-}
-
-function Peers (dispatch, getState) {
-  if (!getState) throw new Error('need getState')
-
-  var self = this
-  self.remotes = {}
-
-  function honorSharingLevel (sharingLevel, type, devices, peerId) {
-    switch (sharingLevel) {
-      case 'LEECH':
-        if (
-          devices.indexOf(peerId) !== -1 ||
-          type !== 'REQUEST_INVENTORY' ||
-          type !== 'REQUEST_SONG' ||
-          type !== 'MULTICAST_SHARING_LEVEL'
-        ) {
-          debug('I am a leech - cannot send:', type)
-          debug('removing songs that are no longer available')
-          actions.RECEIVE_INVENTORY([], peerId)(dispatch, getState)
-          return false
-        }
-        break
-      case 'PRIVATE':
-        if (
-          devices.indexOf(peerId) !== -1 ||
-          type !== 'MULTICAST_SHARING_LEVEL'
-        ) {
-          debug('I am private - cannot send:', type)
-          debug('removing songs that are no longer available')
-          actions.RECEIVE_INVENTORY({}, peerId)(dispatch, getState)
-          return false
-        }
-        break
-      case 'FRIENDS':
-      case 'EVERYONE':
-        break
-      default: return false
-    }
-    return true
-  }
-
-  self.add = (peer, peerId) => {
-    self.remotes[peerId] = peer
-    var buffer = {}
-
-    peer.on('close', (data) => {
-      if (buffer[peerId]) delete buffer[peerId]
-      self.emit('close', peer, peerId)
-    })
-
-    peer.on('data', (data) => {
-      data = data.toString()
-
-      if (data.startsWith('HEADER')) {
-        data = JSON.parse(data.replace('HEADER', ''))
-
-        if (!buffer[peerId]) buffer[peerId] = {}
-
-        buffer[peerId][data.id] = {
-          chunks: data.chunks,
-          data: ''
-        }
-        return
-      }
-
-      data = JSON.parse(data)
-      if (data.index < buffer[peerId][data.id].chunks) {
-        buffer[peerId][data.id].data += data.chunk
-        return
-      }
-
-      buffer[peerId][data.id].data += data.chunk
-      var object = JSON.parse(buffer[peerId][data.id].data)
-
-      delete buffer[peerId][data.id]
-      if (!buffer[peerId]) delete buffer[peerId]
-
-      self.emit('data', object, peerId)
-    })
-  }
-
-  self.remove = (peerId) => {
-    delete self.remotes[peerId]
-  }
-
-  self.send = (object, peerId) => {
-    debug('>> sending to peer', object.type, object, peerId)
-    var sharingLevel = getState().sync.sharingLevel
-    var devices = getState().devices
-
-    if (!honorSharingLevel(sharingLevel, object.type, devices, peerId)) return
-
-    if (!self.remotes[peerId]) {
-      debug('cannot send to offline peer', peerId)
-      return
-    }
-
-    var string = JSON.stringify(object)
-
-    var id = cuid.slug()
-    var send = true
-    var index = 1
-    var size = string.length
-    var chunkSize = 15000 // trial and error
-    var chunks = Math.ceil(size / chunkSize)
-    var offset = 0
-
-    peers.remotes[peerId].write('HEADER' + JSON.stringify({
-      id,
-      chunks
-    }), 'utf8')
-
-    sendChunked()
-    function sendChunked () {
-      do {
-        var chunk = string.slice(offset, offset + chunkSize)
-        var msg = JSON.stringify({
-          id,
-          index: index++,
-          chunk
-        })
-        send = peers.remotes[peerId].write(msg, 'utf8')
-        offset += chunkSize
-      } while (send && offset < size)
-      if (offset < size) peers.remotes[peerId].once('drain', sendChunked)
-    }
-  }
-
-  self.broadcast = (data) => {
-    debug('>> >> broadcasting to ' + Object.keys(self.remotes).length + ' peers', data.type)
-    var sharingLevel = getState().sync.sharingLevel
-    var devices = getState().devices
-
-    for (let peerId in self.remotes) {
-      if (!honorSharingLevel(sharingLevel, data.type, devices, peerId)) continue
-
-      self.send(data, peerId)
     }
   }
 }
